tools/requirements.txt
# Wildflover Tools Dependencies
# Author: Wildflover

Pillow>=10.0.0
requests>=2.31.0
colorama>=0.4.6
